# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API Rest RCS Express",
    "description": "API Rest RCS Express",
    "contact": {
      "name": "rcsxpress",
      "email": "operaciones@rcsexpress.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Ambiente de Desarrollo"
    },
    {
      "url": "https://scen.rcsexpress.com/api/v1/",
      "description": "Ambiente de Producción"
    }
  ],
  "consumes": ["application/json"],
  "produces": ["application/json"],
...
  "paths": {
...
    "/roles": {
      "get": {
        "tags": ["Roles"],
        "summary": "Obtener todos los roles",
        "description": "API para Obtener todos los roles del sistema",
        "parameters": [
          {
            "name": "agencia",
            "in": "header",
            "description": "Codigo de la Agencia",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Roles"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Roles"],
        "summary": "Crear Roles",
        "description": "API para Crear un nuevo Rol",
        "requestBody": {
          "description": "Post object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Roles"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Roles"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "415": {
            "description": "MediaType No Soportado"
          },
          "500": {
            "description": "Error Interno del Sistema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/500"
                }
              }
            }
          }
        }
      }
    },
    "/roles/{id}": {
      "get": {
        "tags": ["Roles"],
        "summary": "Obtener Rol por su Id",
        "description": "API para Retornar un Rol por su ID específico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Codigo del Rol",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Roles"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Roles"],
        "summary": "Actualizar Rol",
        "description": "API para Actualizar un Rol por su ID específico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Codigo del Rol",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Put object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Roles"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Roles"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "415": {
            "description": "MediaType No Soportado"
          },
          "500": {
            "description": "Error Interno del Sistema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/500"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Roles"],
        "summary": "Borrar Rol",
        "description": "API para Borrar un Rol por su ID específico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Codigo del Rol",
            "required": true, "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "id": {
                      "$ref": "#/components/schemas/Roles/properties/id"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/standarResponses/400"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Roles": {
        "description": "Modelo de Roles",
        "required": ["id", "descripcion", "cod_agencia"],
        "properties": {
          "id": {
            "type": "number",
            "readOnly": true
          },
          "descripcion": {
            "type": "string"
          },
          "cod_agencia": {
            "type": "number"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}